+++ app/helpers/application_helper.rb
@@ -516,69 +516,6 @@ module ApplicationHelper
     end
   end
 
-  # change the default link renderer for will_paginate
-  def will_paginate(collection_or_options = nil, options = {})
-    if collection_or_options.is_a? Hash
-      options = collection_or_options
-      collection_or_options = nil
-    end
-    unless options[:renderer]
-      options = options.merge renderer: PaginationListLinkRenderer
-    end
-    super(*[collection_or_options, options].compact)
-  end
-
-  include Pagy::Frontend
-  # Cf https://github.com/ddnexus/pagy/blob/master/gem/lib/pagy/frontend.rb
-  def pagy_nav(pagy, id: nil, aria_label: nil, **vars)
-    return nil unless pagy
-
-    # Keep will_paginate behavior of showing nothing if only one page
-    return nil if pagy.series.length <= 1
-
-    id = %( id="#{id}") if id
-    a  = pagy_anchor(pagy, **vars)
-
-    html = %(<h4 class="landmark heading">#{t('pagination.title')}</h4>)
-
-    html << %(<ol#{id} class="pagination actions pagy" role="navigation" aria-label="#{aria_label || t('pagination.aria-label')}">)
-
-    prev_text = t("pagination.previous")
-    prev_a =
-      if (p_prev = pagy.prev)
-        a.call(p_prev, prev_text)
-      else
-        %(<span class="disabled">#{prev_text}</span>)
-      end
-    html << %(<li class="previous">#{prev_a}</li>)
-
-    pagy.series(**vars).each do |item| # series example: [1, :gap, 7, 8, "9", 10, 11, :gap, 36]
-      html << %(<li>)
-      html << case item
-              when Integer
-                a.call(item)
-              when String
-                %(<a role="link" aria-disabled="true" aria-current="page" class="current">#{pagy.label_for(item)}</a>)
-              when :gap
-                %(<span class="gap">#{pagy_t('pagy.gap')}</span>)
-              else
-                raise InternalError, "expected item types in series to be Integer, String or :gap; got #{item.inspect}"
-              end
-      html << %(</li>)
-    end
-
-    next_text = t("pagination.next")
-    next_a =
-      if (p_next = pagy.next)
-        a.call(p_next, next_text)
-      else
-        %(<span class="disabled">#{next_text}</span>)
-      end
-    html << %(<li class="next">#{next_a}</li>)
-
-    html << %(</ol>)
-  end
-
   # spans for nesting a checkbox or radio button inside its label to make custom
   # checkbox or radio designs
   def label_indicator_and_text(text)
